HARMONY-1282: Refactor batching code to new module
